refactor(react-ui): move page components under src/components/ (audit I-3)
Four flagship packages mixed BOTH layout conventions — page-level components
loose at src/ root alongside an existing src/components/ dir holding their
dialogs/forms/columns. Normalize to the canonical react-ui-{module} structure
(checklist 7f): all presentational components live under src/components/.

- privacy (6), parties (4), payments (4), ai (4): 18 *-page.tsx moved into
  src/components/ via git mv (history preserved); barrels re-export from the
  new path (public API unchanged); relative imports + __tests__ paths fixed.
- Role-suffix naming kept (*-page / *-dialog / *-form / *-columns): the suffix
  marks the component's role, the directory marks it presentational — both
  carry information, matching the tier-wide taxonomy under components/.

tsc (all packages), lint and vitest (425 tests on the 4 packages) green.
