+++ packages/@granit/react-ui-payments/src/__tests__/transaction-list-page.test.tsx
@@ -1,7 +1,7 @@
 import { sampleTransactions } from '@granit/react-payments/testing';
 import { screen, waitFor } from '@testing-library/react';
 
-import { TransactionListPage } from '../transaction-list-page';
+import { TransactionListPage } from '../components/transaction-list-page';
 
 import { renderWithProviders } from './test-utils';
