+++ packages/@granit/react-ui-privacy/src/__tests__/privacy-agreements-page.test.tsx
@@ -1,6 +1,6 @@
 import { screen } from '@testing-library/react';
 
-import { PrivacyAgreementsPage } from '../privacy-agreements-page';
+import { PrivacyAgreementsPage } from '../components/privacy-agreements-page';
 
 import { renderWithProviders } from './test-utils';
