+++ packages/@granit/react-ui-privacy/src/__tests__/privacy-deletion-page.test.tsx
@@ -1,6 +1,6 @@
 import { screen } from '@testing-library/react';
 
-import { PrivacyDeletionPage } from '../privacy-deletion-page';
+import { PrivacyDeletionPage } from '../components/privacy-deletion-page';
 
 import { renderWithProviders } from './test-utils';
