+++ packages/@granit/react-ui-privacy/src/__tests__/privacy-opt-out-page.test.tsx
@@ -1,7 +1,7 @@
 import { screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
-import { PrivacyOptOutPage } from '../privacy-opt-out-page';
+import { PrivacyOptOutPage } from '../components/privacy-opt-out-page';
 
 import { renderWithProviders } from './test-utils';
