+++ packages/@granit/react-ui-privacy/src/components/privacy-agreements-page.tsx
