+++ packages/@granit/react-ui-privacy/src/components/privacy-regulation-page.tsx
