+++ packages/@granit/react-ui-privacy/src/index.ts
@@ -1,10 +1,10 @@
 // Pages — privacy self-service + admin DSR
-export { PrivacyExportPage } from './privacy-export-page';
-export { PrivacyDeletionPage } from './privacy-deletion-page';
-export { PrivacyAgreementsPage } from './privacy-agreements-page';
-export { PrivacyRegulationPage } from './privacy-regulation-page';
-export { PrivacyOptOutPage } from './privacy-opt-out-page';
-export { PrivacyAdminDsrPage } from './privacy-admin-dsr-page';
+export { PrivacyExportPage } from './components/privacy-export-page';
+export { PrivacyDeletionPage } from './components/privacy-deletion-page';
+export { PrivacyAgreementsPage } from './components/privacy-agreements-page';
+export { PrivacyRegulationPage } from './components/privacy-regulation-page';
+export { PrivacyOptOutPage } from './components/privacy-opt-out-page';
+export { PrivacyAdminDsrPage } from './components/privacy-admin-dsr-page';
 
 // Pages — legal document admin
 export { LegalDocumentListPage } from './legal-documents/legal-document-list-page';
